+++ src/packages/button/button.taro.tsx
@@ -27,6 +27,7 @@ export interface ButtonProps
   size: ButtonSize
   block: boolean
   icon: string
+  iconSize: string | number
   children: any
   onClick: (e: MouseEvent) => void
 }
@@ -53,6 +54,7 @@ const defaultProps = {
   size: 'normal',
   block: false,
   icon: '',
+  iconSize: '16',
   style: {},
   children: undefined,
   onClick: (e: MouseEvent) => {},
@@ -68,6 +70,7 @@ export const Button: FunctionComponent<Partial<ButtonProps>> = (props) => {
     size,
     block,
     icon,
+    iconSize,
     children,
     onClick,
     className,
@@ -127,6 +130,7 @@ export const Button: FunctionComponent<Partial<ButtonProps>> = (props) => {
     size,
     block,
     icon,
+    iconSize,
     children,
     onClick,
     classes,
@@ -162,6 +166,7 @@ export const Button: FunctionComponent<Partial<ButtonProps>> = (props) => {
             classPrefix={iconClassPrefix}
             fontClassName={iconFontClassName}
             name={icon}
+            size={iconSize}
           />
         ) : (
           ''