+++ src/packages/button/demo.tsx
@@ -16,6 +16,7 @@ interface T {
   '7db1a8b2': string
   a52bef0c: string
   d04fcbda: string
+  '0aaad622': string
   '0aaad620': string
   '60a53514': string
   b8a453e3: string
@@ -41,6 +42,7 @@ const ButtonDemo = () => {
       '7db1a8b2': '禁用状态',
       a52bef0c: '加载状态',
       d04fcbda: '加载中',
+      '0aaad622': '图标按钮',
       '0aaad620': '按钮尺寸',
       '60a53514': '收藏',
       b8a453e3: '大号按钮',
@@ -63,6 +65,7 @@ const ButtonDemo = () => {
       '7db1a8b2': '禁用狀態',
       a52bef0c: '載入狀態',
       d04fcbda: '載入中',
+      '0aaad622': '图标按钮',
       '0aaad620': '按鈕尺寸',
       '60a53514': '收藏',
       b8a453e3: '大號按鈕',
@@ -85,6 +88,7 @@ const ButtonDemo = () => {
       '7db1a8b2': 'Disabled State',
       a52bef0c: 'Load State',
       d04fcbda: 'Loading',
+      '0aaad622': 'Icon Button',
       '0aaad620': 'Button size',
       '60a53514': 'Favorite',
       b8a453e3: 'Large button',
@@ -130,10 +134,10 @@ const ButtonDemo = () => {
 
         <h2>{translated.e51e4582}</h2>
         <Cell className="button-cell">
-          <Button plain style={{ margin: 8 }} type="primary">
+          <Button plain style={{ margin: 8 }}>
             {translated.e51e4582}
           </Button>
-          <Button plain style={{ margin: 8 }} type="info">
+          <Button plain color="red" style={{ margin: 8 }}>
             {translated.e51e4582}
           </Button>
         </Cell>
@@ -171,7 +175,7 @@ const ButtonDemo = () => {
           </Button>
         </Cell>
 
-        <h2>{translated['0aaad620']}</h2>
+        <h2>{translated['0aaad622']}</h2>
 
         <Cell className="button-cell">
           <Button
@@ -185,6 +189,16 @@ const ButtonDemo = () => {
             shape="square"
             type="primary"
             icon="star"
+            style={{ margin: 8 }}
+          >
+            {translated['60a53514']}
+          </Button>
+          <Button
+            shape="round"
+            type="primary"
+            size="large"
+            icon="star"
+            iconSize={20}
             style={{ margin: 8 }}
           >
             {translated['60a53514']}