+++ src/packages/form/form.tsx
@@ -28,20 +28,28 @@ const PositionInfo: any = {
 
 export const Form: FunctionComponent<
   Partial<FormProps> & React.HTMLAttributes<HTMLFormElement>
-> & { Item: typeof FormItem } = (props) => {
+> & { Item: typeof FormItem } & { useForm: typeof useForm } = (props) => {
   const {
     children,
     onFinish,
     onFinishFailed,
     labelPosition,
     starPositon,
+    form,
     ...rest
   } = {
     ...defaultProps,
     ...props,
   }
 
-  const [formInstance] = useForm()
+  let formInstance: any = {}
+
+  if (Object.keys(form).length !== 0) {
+    formInstance = form
+  } else {
+    ;[formInstance] = useForm()
+  }
+
   formInstance.starPositon = starPositon
   const { setCallback, submit } = formInstance
 
@@ -72,3 +80,4 @@ export const Form: FunctionComponent<
 Form.defaultProps = defaultProps
 Form.displayName = 'NutForm'
 Form.Item = FormItem
+Form.useForm = useForm