+++ src/packages/popover/popover.taro.tsx
@@ -5,14 +5,13 @@ import React, {
   useRef,
   useState,
 } from 'react'
+import { ITouchEvent } from '@tarojs/components'
 import Trigger from './Trigger'
 import Icon from '@/packages/icon/index.taro'
 import Overlay from '@/packages/overlay/index.taro'
 import { getRectByTaro } from '../../utils/useClientRect'
 
 import { BasicComponent, ComponentDefaults } from '@/utils/typings'
-
-import { ITouchEvent } from '@tarojs/components'
 
 export type PopoverTheme = 'light' | 'dark'
 
@@ -164,31 +163,37 @@ export const Popover: FunctionComponent<
               <div className={`${popoverContent}`} style={getStyle()}>
                 <div className={`${popoverArrow}`} />
                 {Array.isArray(children) ? children[1] : ''}
-                {list.map((item: List, i: number) => {
-                  return (
-                    <div
-                      key={item.name}
-                      className={`popover-menu-item ${
-                        item.disabled ? 'disabled' : ''
-                      }`}
-                      onClick={() => {
-                        handleChoose(item, i)
-                      }}
-                    >
-                      {item.icon ? (
-                        <Icon
-                          classPrefix={iconClassPrefix}
-                          fontClassName={iconFontClassName}
-                          className="popover-menu-item-img"
-                          name={item.icon}
-                        />
-                      ) : (
-                        ''
-                      )}
-                      <div className="popover-menu-item-name">{item.name}</div>
-                    </div>
-                  )
-                })}
+
+                <div>
+                  {' '}
+                  {list.map((item: List, i: number) => {
+                    return (
+                      <div
+                        key={item.name}
+                        className={`popover-menu-item ${
+                          item.disabled ? 'disabled' : ''
+                        }`}
+                        onClick={() => {
+                          handleChoose(item, i)
+                        }}
+                      >
+                        {item.icon ? (
+                          <Icon
+                            classPrefix={iconClassPrefix}
+                            fontClassName={iconFontClassName}
+                            className="popover-menu-item-img"
+                            name={item.icon}
+                          />
+                        ) : (
+                          ''
+                        )}
+                        <div className="popover-menu-item-name">
+                          {item.name}
+                        </div>
+                      </div>
+                    )
+                  })}
+                </div>
               </div>
             ) : null}
           </div>