+++ src/packages/progress/demo.taro.tsx
@@ -1,4 +1,5 @@
 import React, { useState } from 'react'
+import Taro from '@tarojs/taro'
 import { useTranslate } from '@/sites/assets/locale/taro'
 import {
   Cell,
@@ -8,7 +9,6 @@ import {
   Toast,
 } from '@/packages/nutui.react.taro'
 import Header from '@/sites/components/header'
-import Taro from '@tarojs/taro'
 
 interface T {
   basic: string
@@ -89,8 +89,9 @@ const ProgressDemo = () => {
         <Cell style={cellStyles}>
           <Progress
             percentage={30}
-            strokeColor="rgba(250,44,25,0.47)"
-            stroke-width="20"
+            fillColor="rgba(250,44,25,0.2)"
+            strokeColor="rgba(250,44,25,0.9)"
+            strokeWidth="15"
             textColor="red"
           />
         </Cell>