+++ src/packages/radio/demo.tsx
@@ -95,6 +95,8 @@ const RadioDemo = () => {
             <Radio.RadioGroup value="1">
               <Radio value="1">{translated.bb7486f4}1</Radio>
               <Radio disabled value="2">
+                {translated.bb7486f4}
+                {translated.bb7486f4}
                 {translated.bb7486f4}2
               </Radio>
               <Radio value="3">{translated.bb7486f4}3</Radio>
@@ -128,7 +130,13 @@ const RadioDemo = () => {
           <Cell>
             <Radio.RadioGroup value="1" direction="horizontal">
               <Radio value="1">{translated.bb7486f4}1</Radio>
-              <Radio value="2">{translated.bb7486f4}2</Radio>
+              <Radio value="2">
+                {translated.bb7486f4}
+                {translated.bb7486f4}
+                {translated.bb7486f4}
+                {translated.bb7486f4}
+                {translated.bb7486f4}2
+              </Radio>
               <Radio value="3">{translated.bb7486f4}3</Radio>
             </Radio.RadioGroup>
           </Cell>
@@ -161,8 +169,8 @@ const RadioDemo = () => {
         <CellGroup>
           <Cell>
             <Radio.RadioGroup value="2">
-              <Radio value="1" iconSize="24">
-                {translated['8a2e2847']}24
+              <Radio value="1" iconSize="15">
+                {translated['8a2e2847']}15
               </Radio>
               <Radio value="2" iconSize="12">
                 {translated['8a2e2847']}12
@@ -178,7 +186,10 @@ const RadioDemo = () => {
                 {translated['70ffa5d8']}
               </Radio>
               <Radio value="2" iconName="checklist" iconActiveName="checklist">
-                {translated['70ffa5d8']}
+                <div>{translated['70ffa5d8']}</div>
+                <div style={{ fontSize: '12px', color: '#8c8c8c' }}>
+                  {translated['70ffa5d8']}
+                </div>
               </Radio>
             </Radio.RadioGroup>
           </Cell>
@@ -203,7 +214,7 @@ const RadioDemo = () => {
               options={optionsDemo1}
               value={radioVal1}
               onChange={(e) => handleChange1(e)}
-            ></Radio.RadioGroup>
+            />
           </Cell>
         </CellGroup>
       </div>