+++ src/packages/searchbar/searchbar.taro.tsx
@@ -1,5 +1,5 @@
 import React, { FunctionComponent, useEffect, useRef, useState } from 'react'
-import Taro from '@tarojs/taro'
+import { getEnv } from '@tarojs/taro'
 import bem from '@/utils/bem'
 import { useConfig } from '@/packages/configprovider/configprovider.taro'
 import Icon from '@/packages/icon/index.taro'
@@ -22,6 +22,7 @@ export interface SearchBarProps extends BasicComponent {
   maxLength?: number
   /** 是否启用清除图标，点击清除图标后会清空输入框	 */
   clearable?: boolean
+  clearIconSize?: string | number
   /** 搜索框外部背景色	 */
   background?: string
   /** 搜索框背景色	 */
@@ -75,6 +76,7 @@ const defaultProps = {
   disabled: false,
   maxLength: 9999,
   clearable: true,
+  clearIconSize: '12px',
   align: 'left',
   readonly: true,
   autoFocus: false,
@@ -101,6 +103,7 @@ export const SearchBar: FunctionComponent<
     disabled,
     maxLength,
     clearable,
+    clearIconSize,
     align,
     readOnly,
     autoFocus,
@@ -164,7 +167,7 @@ export const SearchBar: FunctionComponent<
           shape === 'round' ? searchbarBem('round') : ''
         } ${clearable ? searchbarBem('input-clear') : ''}`}
         ref={searchRef}
-        style={{ ...props.style, background: props.inputBackground }}
+        style={{ ...props.style }}
         value={value || ''}
         placeholder={placeholder || locale.placeholder}
         disabled={disabled}
@@ -253,7 +256,7 @@ export const SearchBar: FunctionComponent<
           classPrefix={iconClassPrefix}
           fontClassName={iconFontClassName}
           name="circle-close"
-          size="12"
+          size={clearIconSize}
           color="#555"
         />
       </div>
@@ -303,7 +306,7 @@ export const SearchBar: FunctionComponent<
   }
 
   const envClass = () => {
-    return Taro.getEnv() === 'WEB' ? 'nut-searchbar-taro' : ''
+    return getEnv() === 'WEB' ? 'nut-searchbar-taro' : ''
   }
 
   return (
@@ -315,7 +318,10 @@ export const SearchBar: FunctionComponent<
     >
       {renderLeftoutIcon()}
       {renderLabel()}
-      <div className={`${searchbarBem('content')}`}>
+      <div
+        className={`${searchbarBem('content')}`}
+        style={{ background: props.inputBackground }}
+      >
         {renderLeftinIcon()}
         <div className="nut-searchbar__input-box">{renderField()}</div>
         {renderRightinIcon()}